requirements.txt
numpy>=1.21.0
pyboy>=2.5.1       # Game Boy emulator
pygame>=2.1.0       # For display and input handling
pillow>=9.0.0      # Image processing
pydantic>=2.0.0    # Data validation
pytest>=7.0.0      # Testing
tqdm>=4.65.0       # Progress bars
rich>=13.0.0       # Rich terminal output
gymnasium>=0.29.0   # For RL environment compatibility 
opencv-python # for screen recording
openai
anthropic
ImageHash # for image hashing
# Core dependencies
playwright>=1.40.0
litellm>=1.0.0
asyncio>=3.4.3

# Optional dependencies
python-dotenv>=1.0.0

# For GUI
sv_ttk